match/pyproject.toml
[tool.poetry]
name = "match"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.7.2"
sqlparse = "0.4.4"
tzdata = "2024.1"
django = "4.2.11"
pandas = "^2.2.1"
numpyencoder = "^0.3.0"
openai = "^1.14.3"
python-dotenv = "^1.0.1"
torch = "^2.2.2"
tqdm = "^4.66.2"
numpy = "^1.26.4"
ipykernel = "^6.29.4"
transformers = "^4.39.1"
scipy = "^1.12.0"
wheel = "^0.43.0"
fasttext-wheel = "^0.9.2"
scikit-learn = "^1.4.1.post1"
haversine = "^2.8.1"
kiwipiepy = "^0.17.0"
matplotlib = "^3.8.3"
selenium = "^4.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
